// src/model/assignmentDetail.ts
/**
 * Grader Extension API Schemas
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { AssignmentSettings } from './assignmentSettings';
import { Submission } from './submission';


export interface AssignmentDetail { 
    id?: number;
    name?: string;
    settings?: AssignmentSettings;
    status?: AssignmentDetail.StatusEnum;
    points?: number;
    submissions?: Array<Submission>;
}
export namespace AssignmentDetail {
    export type StatusEnum = 'created' | 'pushed' | 'released' | 'complete';
    export const StatusEnum = {
        Created: 'created' as StatusEnum,
        Pushed: 'pushed' as StatusEnum,
        Released: 'released' as StatusEnum,
        Complete: 'complete' as StatusEnum
    };
}